+++ packages/backend/jest.config.js
@@ -1,0 +1,17 @@
+const { createDefaultPreset } = require("ts-jest");
+
+const tsJestTransformCfg = createDefaultPreset().transform;
+
+/** @type {import("jest").Config} **/
+module.exports = {
+  testEnvironment: "node",
+
+  // Coverage
+  collectCoverage: true,
+  coverageDirectory: "coverage",
+  coverageProvider: "v8",
+
+  transform: {
+    ...tsJestTransformCfg,
+  },
+};